+++ CHANGELOG.md
@@ -14,6 +14,10 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 - Declarative configuration through `package.json` / `cds-rc.json`
 - Programmatic registration is still possible, but disables auto mode then
 
+### Changed
+
+- Make dependency to `@sap/cds-dk` optional.  If running with the `cds` executable during development, the dependency is fulfilled anyways. Only when running as part of a deployed app with `cds-serve`, then such a dependency is needed.
+
 ## Version 0.7.0 - 2023-12-04
 
 ### Changed